requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.8.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.32.5
playwright>=1.40.0
apscheduler>=3.10.0